% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.r
\name{disp}
\alias{disp}
\title{Calculate the longitudinal dispersivity value based on the approach developed by Xu and Eckstein (1995)}
\usage{
disp(PLUME_L = 1, UNITS = "feet")
}
\arguments{
\item{PLUME_L}{is the plume length}

\item{UNITS}{is the length unit for plume length and longitudinal dispersivity}
}
\description{
This function takes plume length (PLUME_L) and length units (UNITS) as arguments to calculate 
longitudinal dispersivity (α_L). Length units can be meters or feet. The function will return the 
longitudinal dispersivity in the same length units as plume length.
}
\details{
disp(PLUME_L, UNITS)
}
\examples{
disp(PLUME_L = 880, UNITS = "feet")
[1] 23.18827
}
