% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dims.r
\name{get_dims}
\alias{get_dims}
\title{Get the dimensions of a MODFLOW Model without reading the entire .dis file}
\usage{
get_dims(rootname)
}
\arguments{
\item{rootname}{This is the root name of the dis file}
}
\description{
This function is used for extracting the number of rows, columns, layers, stress periods of a model
without reading the entire discritization file
It can be used as a utility function for the other MFtools functions
}
\examples{
get_dims(rootname)
# A tibble: 1 x 4
   NLAY  NROW  NCOL  NPER
  <int> <int> <int> <int>
1    34   328   291     1

D <- get_dims(rootname)
D$NLAY
[1] 34
D$NCOL
[1] 291
D$NLAY * D$NCOL
[1] 9894
}
