% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getroot.r
\name{getroot}
\alias{getroot}
\title{Get the rootname of a MODFLOW Model in a working directory}
\usage{
getroot()
}
\description{
This function is used for extracting the rootname of a MODFLOW model
It can be used as a utility function for the other MFtools functions
}
\examples{
rtnm <- getroot()
rtnm
[1] "F78RM"

}
