% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggres.r
\name{ggres}
\alias{ggres}
\title{Plot the Observed Versus Simulated plot based on 
the PEST .res file}
\usage{
ggres(PATH = NA)
}
\arguments{
\item{PATH}{This is the full file path to the .res file. 
When PATH is NA this function will look in the current working 
directory for the .res file.}
}
\description{
This function reads in the PEST .res file and generates an Observed
Versus Simulated plot
}
