% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfplotveff.R
\name{mfplotveff}
\alias{mfplotveff}
\title{Plot effective volume}
\usage{
mfplotveff(mf)
}
\arguments{
\item{mf}{List produced by \code{\link{mffit}}}
}
\description{
This function plots the function \code{Veff(x)} used for the mass function fit, as stored in the sublist \code{fit} of the output produced by \code{\link{mffit}}.
}
\author{
Danail Obreschkow
}
\seealso{
See examples in \code{\link{mffit}}.
}

