% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfplotvmax.R
\name{mfplotvmax}
\alias{mfplotvmax}
\title{Plot effective volume}
\usage{
mfplotvmax(mf)
}
\arguments{
\item{mf}{List produced by \code{\link{mffit}}}
}
\description{
This function plots the function \code{Vmax(x)} used for the mass function fit, as stored in the sublist \code{fit} of the list produced by \code{\link{mffit}}.
}
\examples{
data = mfdata()
mf = mffit(data$x, data$vmax, data$xerr, write.fit = FALSE)
mfplotvmax(mf)

}
\author{
Danail Obreschkow, 2017
}
\seealso{
\code{\link{mffit}}
}

