% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mftools.R
\docType{package}
\name{mftools-package}
\alias{mftools}
\alias{mftools-package}
\title{Fitting mass and luminosity functions}
\description{
\pkg{mftools} optimally fits galaxy mass functions (MFs) to a set of astronomical data points with/without measurement uncertainties. No binning of the data is used in the fitting. The package can also be used to fit related functions such as halo MFs and luminosity functions.
}
\details{
The functions you're most likely to need from \pkg{mftools} are \code{\link{mffit}} and
\code{\link{mfplot}}. See the documentation of these functions for details.
}

