% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfwrite.R
\name{mfwrite}
\alias{mfwrite}
\title{Write best fitting parameters}
\usage{
mfwrite(mf)
}
\arguments{
\item{mf}{List produced by \code{\link{mffit}}}
}
\description{
This function displays the best-fitting model parameters of a mass function previously fitted using \code{\link{mffit}}.
}
\examples{
data = mfdata()
mf = mffit(data$mass, data$veff, write.fit = FALSE)
mfwrite(mf)

}
\author{
Danail Obreschkow
}
\seealso{
See examples in \code{\link{mffit}}.
}

