% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_MF_CELL_INDX.r
\name{parse_MF_CELL_INDX}
\alias{parse_MF_CELL_INDX}
\title{Parse the elements of a line of a fixed width MODFLOW file}
\usage{
parse_MF_CELL_INDX(INDX = CELLINDX, LINE = inline)
}
\arguments{
\item{LINE}{this is the individual fixed width MODFLOW file line
from the MODFLOW fixed width file}
}
\description{
This function is used for extracting all of the elements of a single line from a fixed width MODFLOW file
It is being used as a utility function for the other MFtools functions
}
