% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_MF_FW_ELMT.r
\name{parse_MF_FW_ELMT}
\alias{parse_MF_FW_ELMT}
\title{Parse the elements of a line of a fixed width MODFLOW file}
\usage{
parse_MF_FW_ELMT(LINE = inline, ELEMENT = 1)
}
\arguments{
\item{LINE}{this is the individual fixed width MODFLOW file line}

\item{ELEMENT}{this is the integer number of the element the user wishes to select
from the MODFLOW fixed width file}
}
\description{
This function is used for extracting individual elements of a single line from a fixed width MODFLOW file
It is being used as a utility function for the other MFtools functions
}
\examples{
infl <- paste0(rootname, ".dis")
linin <- read_lines(infl) \%>\% .[-grep("#", .)]              # READ IN DIS FILE BUT REMOVE COMMENTED LINES
                                                            # THIS IS IN PREPARATION FOR MODFLOW 6
indx <- 1
NLAY <- linin[indx] \%>\% parse_MF_FW_ELMT(1) \%>\% as.integer()
NLAY
[1] 15
}
