% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_MF_FW_LINE.r
\name{parse_MF_FW_LINE}
\alias{parse_MF_FW_LINE}
\title{Parse the elements of a line of a fixed width MODFLOW file}
\usage{
parse_MF_FW_LINE(LINE = inline)
}
\arguments{
\item{LINE}{this is the individual fixed width MODFLOW file line
from the MODFLOW fixed width file}
}
\description{
This function is used for extracting all of the elements of a single line from a fixed width MODFLOW file
It is being used as a utility function for the other MFtools functions
}
\examples{
dY <- linin[indx + seq(1:BLOCKEND) - 1] \%>\% parse_MF_FW_LINE()
}
