% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltpstbarbell.r
\name{pltpstbarbell}
\alias{pltpstbarbell}
\title{Plot a barbell plot of the observed (blue) and simulated (red) groundwater elevations as listed
in the pest .res file. This function requires that both a .res file and targpest.dat file are
located in the PATH directory.}
\usage{
pltpstbarbell(PATH = NA, MAIN = NA, colour = "#a3c4dc",
  colour_obs = "blue", colour_sim = "red", size_obs = 3,
  size_sim = 3, alpha = 0.6, size = 1, xlab = "", ylab = "")
}
\arguments{
\item{PATH}{This is the full file path to the .res and targpest.dat files. They need to be located in the same directory.}

\item{MAIN}{This is text to be printed as the main figure title}

\item{colour}{the colour of the bar}

\item{colour_obs}{the colour of the observed groundwater elevation}

\item{colour_sim}{the colour of the simulated groundwater elevation}

\item{size_obs}{the size of the observed groundwater elevation}

\item{size_sim}{the size of the simulated groundwater elevation}

\item{alpha}{the level of transparency of the ends}

\item{size}{the size of the bar}
}
\description{
Plot a barbell plot of the observed (blue) and simulated (red) groundwater elevations as listed
in the pest .res file. This function requires that both a .res file and targpest.dat file are
located in the PATH directory.
}
