% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstmodcal.r
\name{pstmodcal}
\alias{pstmodcal}
\title{Report model diagnostic statistics based on the PEST .res file.
This function requires a PEST .res file to be located in the PATH directory.}
\usage{
pstmodcal(PATH)
}
\arguments{
\item{PATH}{this is the full file path to the .res file.}
}
\description{
Report model diagnostic statistics based on the PEST .res file.
This function requires a PEST .res file to be located in the PATH directory.
}
