% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmg_lst.r
\name{read_gmg_lst}
\alias{read_gmg_lst}
\title{Extract Head-Change Values and Cell Locations from the .lst file}
\usage{
read_gmg_lst(LISTFL)
}
\arguments{
\item{LISTFL}{This is the full name (including filepath) of the .lst file.}
}
\description{
This function extracts the Head-Change value and the locations of the cells exhibiting the
maximum head-change value from the .lst file.
}
