% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nwt_lst.r, R/read_pcg_lst.r
\name{read_pcg_lst}
\alias{read_pcg_lst}
\alias{read_pcg_lst}
\title{Extract Head-Change Values and Cell Locations from the .lst file}
\usage{
read_pcg_lst(LISTFL)

read_pcg_lst(LISTFL)
}
\arguments{
\item{LISTFL}{This is the full name (including filepath) of the .lst file.}

\item{LISTFL}{This is the full name (including filepath) of the .lst file.}
}
\description{
This function extracts the Head-Change value and the locations of the cells exhibiting the
maximum head-change value from the .lst file.

This function extracts the Head-Change value and the locations of the cells exhibiting the
maximum head-change value from the .lst file.
}
