% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbtn.r
\name{readbtn}
\alias{readbtn}
\title{Read MT3D .btn File}
\usage{
readbtn(rootname)
}
\arguments{
\item{rootname}{This is the root name of the btn file}
}
\description{
This function reads in a btn file and creates a list
composed of the following vectors:
\describe{
\item{NLAY}{Atomic Vector of the Number of Layers in the Model Grid}
\item{NCOL}{Atomic Vector of the Number of Columns in the Model Grid} 
\item{NROW}{Atomic Vector of the Number of Rows in the Model Grid} 
\item{NPER}{Atomic Vector of the Number of Stress Periods in the Model}
\item{NCOMP}{Atomic Vector of the Number of Compounds in the Model} 
\item{MCOMP}{Atomic Vector of the Number of Mobile Compounds in the Model} 
\item{TUNIT}{Atomic Vector of the Number of Time Unit}
\item{LUNIT}{Atomic Vector of the Number of Length Unit}
\item{MUNIT}{Atomic Vector of the Number of Mass Unit} 
\item{TRNOP}{Logical Vector of Packages To Use}
\item{LAYCON}{Vector of Layer Types: LAYCON = 0 = Confined, LAYCON != 0 = Unconfined or Convertible} 
\item{dX}{Cell Width Along Columns} 
\item{X}{Cell Center Coordinate in the Model Coordinate System} 
\item{dY}{Cell Width Along Rows}
\item{Y}{Cell Center Coordinate in the Model Coordinate System}
\item{TOP}{Top Elevation of Layer 1}
\item{TRANS} {Data frame composed of LAY, ROW, COL, dZ, PRSITY, ICBUND, SONC, where these values are defined as follows:}
\item{dZ}{Layer Thicknesses}
\item{PRSITY}{Porosity} 
\item{ICBUND}{ICBUND array. See MT3DMS manual for a description of ICBUND}
\item{SCONC}{Starting Concentrations at the beginning of the simulation} 
\item{CINACT}{Value that indicates an inactive value} 
\item{THKMIN}{Minimum saturated thickness in a cell}
\item{IFMTCN}{Flag indicating whether the calculated concentration should be 
printed to the standard output text file and also serves as a printing-format
code if it is printed}
\item{IFMTNP}{Flag indicating if the number of particles in each cell should be printed and also
serves as a printing format code if it is printed}
\item{IFMTRF}{Flag indicating if the model-calculated retardation factor should be printed and also
serves as a printing format code if it is printed}
\item{IFMTDP}{Flag indicating if the model-calculated distance-weighted dispersion coefficient should
be printed and also serves as a printing format code if it is printed} 
\item{SAVUCN}{Logical flag indicating if the concentration solution should be saved to a .ucn file}
\item{NPRS}{Flag indicating the frequncy of output and also indicating whether the frequency output
is specified in terms of total elapsed simulation time or the transport step number.
See MT3D Manual for further details of NPRS}
\item{TIMPRS}{Total elapsed time at which the simulation results are printed}
\item{NOBS}{Number of Observations wells recording concentration data} 
\item{NPROBS}{Integer indicating how frequently the concentration is recorded at each
 Observation Well} 
\item{OBSLOC}{Cell indices (L, R, C) of Observation Well locations} 
\item{CHKMAS}{Logical Flag indicating whether a one-line summary of mass balance
information should be printed} 
\item{NPRMAS}{Integer indicating how frequently the mass budget information
should be recorded} 
\item{PERLEN}{Lengths of Stress Perids}
\item{NSTP}{Number of time-steps for the transient flow folution in each stress-period} 
\item{TSMULT}{Time-step multiplier. Only used if NSTP > 1. See MT3DMS Manual for 
further details} 
\item{TSLNGHdf}{If TSMULT < 0, TSLNGHdf frovides the length of time-steps for the
flow solution in each stress period}
\item{DT0}{The user-specified transport step size within each time-step of the
flow solution} 
\item{MXSTRN}{The maximum number of transport time steps allowed within one time
step of the flow solution}
\item{TTSMULT}{A multiplier for successive transport steps within a flow step} 
\item{TTSMAX}{Maximum transport step size allowed when transport step size multiplier
TTSMULT > 1.0}
}
}
\examples{
readbtn("T04")

# Find the average initial concentration assigned to the MT3DMS Model
btn <- readbtn("T04")
btn$TRANS \%>\% 
select(SCONC) \%>\%
summarise(AVG = mean(SCONC))

> A tibble: 1 x 1
>        AVG
>      <dbl>
> 1 17.77554

# Find the average initial concentration (> 5 µg/L) 
# assigned to Layer 1 of the MT3DMS Model
btn$TRANS \%>\% 
filter(LAY == 1) \%>\%
select(SCONC) \%>\%
filter(SCONC >= 5) \%>\%
summarise(AVG = mean(SCONC))

> A tibble: 1 x 1
>        AVG
>      <dbl>
> 1 1096.754

# Find the range of porosity values assigned to the model 
btn$TRANS \%>\%                  # Select the RCL Value table that contains the porosity values
select(PRSITY) \%>\%             # Select the table column that contains porosity values
t() \%>\%                        # Convert to a vector of values
as.factor() \%>\%
levels()

> [1] "0.15" "0.21" "0.25" "0.4" 
}
