% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdis.r
\name{readdis}
\alias{readdis}
\title{Read MODFLOW .dis File}
\usage{
readdis(rootname = NA)
}
\arguments{
\item{rootname}{This is the root name of the dis file}
}
\description{
This function reads in a dis file and creates a list
composed of the following vectors:
\describe{
\item{NLAY}{Atomic Vector of the Number of Layers in the Model Grid}
\item{NROW}{Atomic Vector of the Number of Rows in the Model Grid}
\item{NCOL}{Atomic Vector of the Number of Columns in the Model Grid}
\item{NPER}{Atomic Vector of the Number of Stress Periods in the Model Grid}
\item{ITMUNI}{Atomic Vector Indicating the Time Unit of Model Data : 0 = Undefined, 1 = Seconds, 2 = Minutes, 3 = Hours, 4 = Days, 5 = Years}
\item{LENUNI}{Atomic Vector Indicating the Length Unit of Model Data : 0 = Undefined, 1 = Feet, 2 = Meters, 3 = Centimeters}
\item{LAYCBD}{Flag With One Value for Each Model Layer that Indicates Whether or Not a Layer has a Quasi-3D Confining Bed Below it. 0 Indicates No Confining Bed and Not 0 Indicates Confining Bed}
\item{dX}{Cell Width Along Columns}
\item{dY}{Cell Width Along Rows}
\item{TOP}{Top Elevation of Layer 1}
\item{BOT}{Bottom Elevation of Each Model Cell or a Quasi-3D Confining Bed}
\item{PERLEN}{Length of Each Stress Period}
\item{NSTP}{Number of Time Steps in a Stress Period}
\item{TSMULT}{Multiplier Length of Successive Time Steps. See the MODFLOW2005 Manual For a Description.}
\item{SS}{Character Vector Indicating Whether The Stressperiod is Transient ("TR") or Steady State ("SS")}
}
}
\examples{
readdis("F95")
$NLAY
[1] 8

$NCOL
[1] 610

$NROW
[1] 414

$NPER
[1] 16

$ITMUNI
[1] 4

$LENUNI
[1] 1

$LAYCBD
  L1 L2 L3 L4 L5 L6 L7 L8
  0  0  0  0  0  0  0  0
etc.
}
