% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readhds.r
\name{readhds}
\alias{readhds}
\title{Read MODFLOW .hds File}
\usage{
readhds(rootname, NLAY, NPER)
}
\arguments{
\item{rootname}{This is the rootname of the flow model (i.e., "F95")}

\item{NLAY}{This is the number of layers assigned to the model and needs to be 
assigned}

\item{NPER}{This is the number of stress periods that are printed to the .hds file. This number
can be obtained from the .dis file}
}
\description{
This function reads in a binary hds file and creates a data frame
composed of the following columns:
\describe{
\item{TRANS}{Transport Time Step}
\item{STP}{Flow Time Step}
\item{PERLEN}{Stress Period Length}
\item{TIME}{Elapsed Time}
\item{LAY}{Model Layer}
\item{ROW}{Model Row}
\item{COL}{Model Column}
\item{GWE}{Groundwater Elevation}
}
}
\examples{
readhds("F95", NLAY = 8, NTTS = 16)
}
