% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readlpf.r
\name{readlpf}
\alias{readlpf}
\title{Read MODFLOW .lpf File}
\usage{
readlpf(rootname)
}
\arguments{
\item{rootname}{This is the root name of the dis file}
}
\description{
This function reads in a lpf file and creates a list
composed of the following vectors:
\describe{
\item{ILPFCB}{is a flag and a unit number. If ILPFCB > 0, cell-by-cell flow terms will be written to this unit
number when "SAVE BUDGET" or a non-zero value for ICBCFL is specified in Output Control. The terms that are saved
are storage, constant-head flow, and flow between adjacent cells.
If ILPFCB = 0, cell-by-cell flow terms will not be written.
If ILPFCB < 0, cell-by-cell flow for constant-head cells will be written in the listing file when "SAVE BUDGET"
or a non-zero value for ICBCFL is specified in Output Control. Cell-by-cell flow to storage and between
adjacent cells will not be written to any file.}
\item{HDRY}{is the head that is assigned to cells that are converted to dry during a simulation. Although this value plays
no role in the model calculations, HDRY values are useful as indicators when looking at the resulting heads that are
output from the model. HDRY is thus similar to HNOFLO in the Basic Package, which is the value assigned to cells
that are no-flow cells at the start of a model simulation.}
\item{NPLPF}{is the number of LPF parameters}
\item{LAYTYP}{contains a flag for each layer that specifies the layer type.
0 – confined
>0 – convertible
<0 – convertible unless the THICKSTRT option is in effect. When THICKSTRT is in effect, a negative value of
LAYTYP indicates that the layer is confined, and its saturated thickness will be computed as STRT-BOT.}
\item{LAYAVG}{contains a flag for each layer that defines the method of calculating interblock transmissivity.
0—harmonic mean
1—logarithmic mean
2—arithmetic mean of saturated thickness and logarithmic-mean hydraulic conductivity.}
\item{CHANI}{contains a value for each layer that is a flag or the horizontal anisotropy. If CHANI is less than or equal to
0, then variable HANI defines horizontal anisotropy. If CHANI is greater than 0, then CHANI is the horizontal
anisotropy for the entire layer, and HANI is not read. If any HANI parameters are used, CHANI for all layers must
be less than or equal to 0.}
\item{LAYVKA}{contains a flag for each layer that indicates whether variable VKA is vertical hydraulic conductivity or
the ratio of horizontal to vertical hydraulic conductivity.
0—indicates VKA is vertical hydraulic conductivity
not 0—indicates VKA is the ratio of horizontal to vertical hydraulic conductivity, where the horizontal hydraulic
conductivity is specified as HK in item 10.}
\item{LAYWET}{contains a flag for each layer that indicates whether wetting is active.
0—indicates wetting is inactive
not 0—indicates wetting is active}
\item{WETFCT}{is a factor that is included in the calculation of the head that is initially established at a cell when the cell
is converted from dry to wet. (See IHDWET.)
IWETIT—is the iteration interval for attempting to wet cells. Wetting is attempted every IWETIT iteration. If using
the PCG solver (Hill, 1990), this applies to outer iterations, not inner iterations. If IWETIT ≤ 0, the value is changed
to 1.}
\item{IHDWET}{is a flag that determines which equation is used to define the initial head at cells that become wet:
If IHDWET = 0, equation 5-32A is used: h = BOT + WETFCT (hn - BOT) .
If IHDWET is not 0, equation 5-32B is used: h = BOT + WETFCT(THRESH)}
\item{PARNAM}{is the name of a parameter to be defined. This name can consist of 1 to 10 characters and is not case
sensitive. That is, any combination of the same characters with different case will be equivalent.}
\item{PARTYP}{is the type of parameter to be defined. For the LPF Package, the allowed parameter types are:
HK—defines variable HK, horizontal hydraulic conductivity
HANI—defines variable HANI, horizontal anisotropy
VK—defines variable VKA for layers for which VKA represents vertical hydraulic conductivity (LAYVKA=0)
VANI—defines variable VKA for layers for which VKA represents vertical anisotropy (LAYVKA≠0)
SS—defines variable Ss, the specific storage
SY—defines variable Sy, the specific yield
VKCB—defines variable VKCB, the vertical hydraulic conductivity of a Quasi-3D confining layer.}
\item{Parval}{is the parameter value. This parameter value may be overridden by a value in the Parameter Value File.}
\item{NCLU}{is the number of clusters required to define the parameter. Each repetition of Item 9 is a cluster (variables
Layer, Mltarr, Zonarr, and IZ). Each layer that is associated with a parameter usually has only one cluster. For
example, parameters which apply to cells in a single layer generally will be defined by just one cluster. However,
having more than one cluster for the same layer is acceptable.}
\item{Layer}{is the layer number to which a cluster definition applies.}
\item{Mltarr}{is the name of the multiplier array to be used to define variable values that are associated with a parameter.
The name “NONE” means that there is no multiplier array, and the variable values will be set equal to Parval.}
\item{Zonarr}{is the name of the zone array to be used to define the cells that are associated with a parameter. The name
“ALL” means that there is no zone array, and all cells in the specified layer are part of the parameter.}
\item{IZ}{is up to 10 zone numbers (separated by spaces) that define the cells that are associated with a parameter. These
values are not used if ZONARR is specified as “ALL”. Values can be positive or negative, but 0 is not allowed. The
end of the line, a zero value, or a non-numeric entry terminates the list of values.}
\item{PROPS}{Data frame of LAY, ROW, COL, HK, HANI, VKA, Ss, Sy, VKCB, WETDRY}
}
}
\examples{
readlpf("F95")
$ILPFCB
[1] 50

$HDRY
[1] -1e+30

$NPLPF
[1] 0

$LAYTYP
[1] 1 3 3 3 0 0 0 0

$LAYAVG
[1] 0 0 0 0 0 0 0 0

$CHANI
[1] -1 -1 -1 -1 -1 -1 -1 -1

etc.

Use this to develop summary statistics of the hydraulic properties
p <- readlpf("F95")
p$PROPS \%>\% group_by(LAY) \%>\% summarise(MIN_K = min(HK), MEDIAN_K = median(HK), MAX_K = max(HK))

}

