% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnoflow.r
\name{readnoflow}
\alias{readnoflow}
\title{Read a no-flow matrix text file}
\usage{
readnoflow(FILE, rootname)
}
\arguments{
\item{FILE}{is the name of the no-flow matrix file}

\item{rootname}{is the rootname of the groundwater flow model.
This will be used to find and read the discretization file}
}
\description{
This function reads a no-flow matrix text file. It first reads the model .dis file to
obtain the dimensions of the finite-diference grid.
}
\details{
readnoflow(FILE, rootname)
}
\examples{
readnoflow(FILE = NF.dat, rootname = "FF_INIT")
}

