% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readpathline.r
\name{readpathline}
\alias{readpathline}
\title{Read the MODPATH pathline file and create a tibble of the contents}
\usage{
readpathline(PATH = NA)
}
\arguments{
\item{PATH}{This is the full file path to the pathline file. 
When PATH is NA this function will look in the current working 
directory for the pathline file.}
}
\description{
This function reads in the MODPATH pathline file and generates a tibble of the contents.
The tibble includes the following columns
\describe{
\item{PART}{Particle Index Number}
\item{GLOBAL_X}{Global coordinate in the x-direction}
\item{GLOBAL_Y}{Global coordinate in the y-direction}
\item{LOCAL_Z}{Local coordinate in the z-direction within the cell}
\item{GLOBAL_Z}{Global coordinate in the z-direction}
\item{CUMULATIVE_TIME}{Cumulative tracking time}
\item{COL}{J index of cell containing the point (i.e., Column Number)}
\item{ROW}{I index of cell containing the point (i.e., Row Number)}
\item{LAY}{K index of cell containing the point (i.e., Layer Number)}
\item{TIME_STEP}{Cumulative MODFLOW time step number}
}
}
