% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readpstres.r
\name{readpstres}
\alias{readpstres}
\title{Read PEST .res file}
\usage{
readpstres(PATH = NA)
}
\arguments{
\item{PATH}{This is the full file path to the .res file. 
When PATH is NA this function will look in the current working 
directory for the .res file.}
}
\description{
This function reads in the PEST .res file
}
