% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readrch.r
\name{readrch}
\alias{readrch}
\title{Read MODFLOW .rch File}
\usage{
readrch(rootname = NA)
}
\arguments{
\item{rootname}{This is the root name of the rch file}
}
\description{
This function reads in a rch file and creates a list
composed of the following vectors:
\describe{
\item{NRCHOP}{Atomic Vector of the is the recharge option code. Recharge fluxes are defined in a layer variable, RECH, with one value for
              each vertical column. Accordingly, recharge is applied to one cell in each vertical column, and the option code
              determines which cell in the column is selected for recharge.
              1—Recharge is only to the top grid layer.
              2—Vertical distribution of recharge is specified in layer variable IRCH.
              3—Recharge is applied to the highest active cell in each vertical column. A constant-head node intercepts
              recharge and prevents deeper infiltration.}
\item{IRCHCB}{Atomic Vector and is flag and a unit number.
              If IRCHCB > 0, cell-by-cell flow terms will be written to this unit number when "SAVE BUDGET" or a nonzero
              value for ICBCFL is specified in Output Control.
              If IRCHCB ≤ 0, cell-by-cell flow terms will not be written.}
\item{RCH}{Data Frame Composed of Layer, Row, Column, Stress Period, and Recharge rate applied to the top of the model domain}
}
}
