% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readucn.r
\name{readucn}
\alias{readucn}
\title{Read MT3DMS .ucn File}
\usage{
readucn(ucrootname, NLAY, NTTS)
}
\arguments{
\item{NLAY}{This is the number of layers assigned to the model}

\item{NTTS}{This is the number of transport time steps that are printed to the .ucn file. This can be obtained from the .btn file. 
This will cause an error if the value assigned is greater than the correct value. Future versions will need to be developed so that NTTS is not
needed for this function to operate properly.}

\item{ucname}{This is the name of the ucn file}
}
\description{
This function reads in a binary ucn file and creates a data frame
composed of the following elements:
\describe{
\item{TRANS}{Transport Time Step}
\item{STP}{Flow Time Step}
\item{PER}{Stress Period}
\item{TIME}{Elapsed Time}
\item{LAY}{Model Layer}
\item{ROW}{Model Row}
\item{COL}{Model Column}
\item{CONC}{Concentration}
}
}
\examples{
readucn("MT3DMS001S", NLAY = 8, NTTS = 22)
}

