% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwel.r
\name{readwel}
\alias{readwel}
\title{Read MODFLOW .wel File}
\usage{
readwel(rootname, NSP = 1)
}
\arguments{
\item{rootname}{This is the root name of the dis file}

\item{NSP}{This is the number of stress periods in the model simulation that will be listed in the well file}
}
\description{
This function reads in a wel file and creates a list
composed of the following vectors:
\describe{
\item{TEXT}{is a character variable (199 characters) that starts in column 2 of the .wel file. Any characters can be included in Text. 
The “#” character is in column 1. Text is written to the Listing File.}
\item{PARAMETER}{The optional items NPWEL and MXL (below) must start with the word "PARAMETER"}
\item{NPWEL}{is the number of well parameters}
\item{MXL}{is the maximum number of wells that will be defined using parameters}
\item{MXACTW}{is the maximum number of wells in use during any stress period, including those that are defined using parameters}
\item{IWELCB}{is a flag and a unit number.
If IWELCB > 0, cell-by-cell flow terms will be written to this unit number when "SAVE BUDGET" or a nonzero
value for ICBCFL is specified in Output Control.
If IWELCB = 0, cell-by-cell flow terms will not be written.
If IWELCB < 0, well recharge for each well will be written to the listing file when "SAVE BUDGET" or a nonzero
value for ICBCFL is specified in Output Control}
\item{Option}{is an optional list of character values.
“AUXILIARY abc” or “AUX abc”—defines an auxiliary variable, named "abc", which will be read for each
well as part of Items 4 and 6. Up to 20 variables can be specified, each of which must be preceded by
"AUXILIARY" or "AUX." These variables will not be used by the Ground-Water Flow Process, but they
will be available for use by other processes. The auxiliary variable values will be read after the Q variable.
“NOPRINT”—specifies that lists of wells will not be written to the Listing File.}
\item{WELL}{A dataframe consisting of columns of Stress-Period (SP), Layer (LAY), Row (ROW), Column (COL), Pumping Rate (Q), and 
auxilliary variable (xyz) for each well cell and each stress-period}
}
}
\examples{
readwl("F10")
$TEXT
[1] "# MODFLOW2000 Well Package"

$PARAMETER
[1] "PARAMETER"

$NPWEL
[1] 0

$MXL
[1] 0

$Option
[1] "   AUX IFACE"

$MXACTW
[1] 105

$IWELCB
[1] 50

$WELL
# A tibble: 13,965 × 6
      SP   LAY   ROW   COL          Q       xyz
   <int> <int> <int> <int>      <dbl>     <chr>
1      1    11   142   135 -13378.750         0
2      1    12   142   135 -13378.750         0
3      1     7   127   243 -10683.750         0
4      1     8   127   243 -10683.750         0
5      1    11    27    46  -9624.999         0
6      1    12    27    46  -9624.999         0
7      1    11    27    47  -7699.999         0
8      1    12    27    47  -7699.999         0
9      1     1   232    39  -2310.000         0
10     1     7   230    94  -9239.999         0
# ... with 13,955 more rows
}

