% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slvrplt.r
\name{slvrplt}
\alias{slvrplt}
\title{Plot and Print Solver Maximum Head-Change Values}
\usage{
slvrplt(PATH = NA, HCLOSE = 0.001, slvUNIT = 19)
}
\arguments{
\item{PATH}{This is the path to the directory that contains model files. The directory needs to include the .in file, .nam file, and .lst file.
If PATH is not specified set the working directory to the directory containing the model files.}

\item{HCLOSE}{This is the head-change close criterion.}

\item{slvUNIT}{This is the solver unit number in the MODFLOW .nam file. This defaults to 19.}
}
\description{
This function reads in the list files and plots the head-change values. It assumes that 
the solver is unit 19 in the .nam file. The slvrplt is currently working for GMG but as the
need arises other solvers will be added. Make sure that the solver is printing all information to the list file.
For the GMG solver this means setting ioutgmg to 2. This function will create a plot showing the head change over successive iterations.
It will also generate a .csv file indicating the cell location of the cell with the maximum head value.
}
