% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targid.r
\name{targid}
\alias{targid}
\title{Link model target identification to PEST target identification as listed in 
Groundwater Vistas targpest.dat file
This function reads the targpest.dat file and gets the target ID based on the
PEST target ID. This assumes that the PEST target ID is in the format of o#, where `#` is the
target number. The first PEST target ID must be o1 and the preceding line in the targpest.dat
file needs to be the number of targets.}
\usage{
targid(PATH = NA)
}
\arguments{
\item{PATH}{This is the full file path to the targpest.dat file. 
When PATH is NA this function will look in the current working 
directory for the targpest.dat file.}
}
\description{
Link model target identification to PEST target identification as listed in 
Groundwater Vistas targpest.dat file
This function reads the targpest.dat file and gets the target ID based on the
PEST target ID. This assumes that the PEST target ID is in the format of o#, where `#` is the
target number. The first PEST target ID must be o1 and the preceding line in the targpest.dat
file needs to be the number of targets.
}
