% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeucn.r
\name{writeucn}
\alias{writeucn}
\title{Write MT3DMS .ucn File}
\usage{
writeucn(ucn, ofl)
}
\arguments{
\item{ucn}{is the ucn data frame}

\item{ofl}{is the output ucn file name}
}
\description{
This function writes a ucn dataframe created in R (i.e., from readucn) to a binary ucn file
composed of the following elements:
}
\details{
writeucn(ucn, ofl)
}
\examples{
writeucn(ucn = u, ofl = "T04R1.ucn")
}
