% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdata.R
\name{mfdata}
\alias{mfdata}
\title{Example data}
\usage{
mfdata()
}
\description{
This function produces an example of log-masses \code{x} with uncertainties \code{xerr} and effective volumes \code{vmax}. The data is taken from a publication by Westmeier et al. (2017) and represents HI-Masses in the Sculptor structure.
}
\examples{
data <- mfdata()
print(data$x)
print(data$vmax)
print(data$xerr)

# These data can then be used to fit a MF, e.g.
mf = mffit(data$x,data$vmax,data$xerr)

}
\author{
Danail Obreschkow, 2017
}
\seealso{
\code{\link{mffit}}
}

