% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdata.R
\name{mfdata}
\alias{mfdata}
\title{Example data}
\usage{
mfdata()
}
\description{
This function produces an example of log-masses \code{x} with Gaussian uncertainties \code{sigma} and effective volumes \code{veff}. The data is taken from a publication by Westmeier et al. (2017) and represents HI-Masses in the Sculptor structure.
}
\examples{
data <- mfdata()
print(data$x)
print(data$veff)
print(data$sigma)

# These data can then be used to fit a MF, e.g.
mf = mffit(data$x, data$veff, data$sigma)

}
\author{
Danail Obreschkow
}
\seealso{
\code{\link{mffit}}
}

