% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mffit.R
\name{mffit}
\alias{mffit}
\title{Fit a galaxy mass function}
\usage{
mffit(x, vmax, x.sigma = NULL, type = "Schechter", niterations = 5,
  resampling = FALSE, resampling.iterations = 100, write.fit = TRUE)
}
\arguments{
\item{x}{Normally a vector of log-masses \code{x = log10(M/Msun)}, but also be used with log-luminosities or absolute magnitudes.}

\item{vmax}{This is either a function or a vector. If vmax is a function, it will be interpreted as the analytical function \code{Veff(x)} representing the effective cosmic volume, in which the log-masses \code{x} can be detected; normally this is an increasing function of \code{x}. If \code{vmax} is a vector of same length as \code{x}, it will be interpreted as the values of the maximum cosmic volumes associated with the different values of \code{x}, and a function \code{Veff(x)} will be fitted automatically. In both cases the continuous function \code{Vmax(x)} will be returned as \code{veff.fn} in the output list.}

\item{x.sigma}{Optional vector of same length as \code{x}, containing the 1-sigma uncertainties of \code{x}.}

\item{type}{Kind of MF: \code{'Schechter'} for Schechter function or \code{'PL'} for a simple power law.}

\item{resampling}{If true, the MF is resampled resampling.iterations times to determine quantiles of the parameter distributions and fitted functions.}

\item{resampling.iterations}{Integer (>0) number of iterations for the resampling. Ignored if \code{resampling = FALSE}.}

\item{write.fit}{Logical argument specifying whether the best-fitting parameters are displayed in the console.}
}
\value{
Returns a structured list. The entry \code{input} contains all the input arguments. The entry \code{fit} contains all the output arguments. Most importantly, the equation of the fitted MF and the best-fitting parameters are contained in the sub-list \code{fit$parameters}.
}
\description{
This function fits an analytical model of a galaxy mass function (MF) to a discrete set of galaxies with observed masses and effective volumes.
}
\examples{
data = mfdata()
mf = mffit(data$x, data$vmax, data$x.sigma, resampling = TRUE)
mfplot(mf, bin.xmin = 6.5, bin.xmax = 9.5, xlim=c(2e6,5e10), ylim=c(1e-3,1))

}
\author{
Danail Obreschkow, 2017
}
\keyword{fit}
\keyword{function}
\keyword{mass}
\keyword{schechter}

