% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfmodel.R
\name{mfmodel}
\alias{mfmodel}
\title{Analytical galaxy mass functions}
\usage{
mfmodel(x = NULL, p = NULL, sigma = NULL, output = "density",
  type = "Schechter")
}
\arguments{
\item{x}{Vector of log-masses, typically \code{x = log10(M/Msun)}.}

\item{p}{Parameters of the analytical function. See argument \code{type}.}

\item{sigma}{Uncertainties of the parameters \code{p}. If sigma is a vector of same length as \code{p}, these uncertainties are considered to be 1-sigma Gaussian uncertainties. If sigma is a 2-row array, the first row contains the upper uncertanties to the 84-percentile value, whereas the second row contains the lower uncertainties down to the 16-percentile value.}

\item{output}{Specifies what the function is doing. \code{'density'} evaluate the MF at \code{x}, \code{'npara'} returns the number of parameters of the analytical function, \code{'default'} returns a vector of typical parameters \code{p} that can be used as initial values for fitting the MF, \code{'equation'} returns a string with the equation of the MF, \code{'parameters'} displays the best fitting parameters with uncertainties.}

\item{type}{Kind of MF: \code{'Schechter'} for Schechter function or \code{'PL'} for a simple power law.}
}
\description{
This function handles and evaluates different models of the galaxy mass function (MF), such as the Schechter function.
}
\examples{
# Evaluate and plot a Schechter function
x = seq(6,10,length=100)
mass = 10^x
parameters = c(-2,9.6,-1.5)
phi = mfmodel(x, parameters, type = 'Schechter')
plot(mass, phi, type='l', log='xy')

# Display the equation and best-fitting parameters wit uncertainties
uncertainties = c(0.3,0.23,0.13)
mfmodel(p = parameters, sigma = uncertainties, output = 'parameters')

}
\author{
Danail Obreschkow, 2017
}
\seealso{
\code{\link{mffit}}
}

