% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfplot.R
\name{mfplot}
\alias{mfplot}
\title{Display fitted galaxy mass function}
\usage{
mfplot(mf, nbins = NULL, bin.xmin = NULL, bin.xmax = NULL,
  xlab = expression("M [M"["sun"] * "]"), ylab = expression(phi ~ "[Mpc"^-3
  ~ "dex"^-1 ~ "]"), xlim = NULL, ylim = NULL,
  show.quantile.50.90 = FALSE)
}
\arguments{
\item{mf}{List produced by \code{\link{mffit}}}

\item{nbins}{Number of bins to be plotted. This is purely for illustrative purposes. The fitting does not use bins. Choose \code{nbins=NULL} (default) to determine the number of bins automatically or \code{nbins=0} to suppress the bins.}

\item{bin.xmin}{Left edge of first bin}

\item{bin.xmax}{Right edge of last bin}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{xlim}{2-element vector with x-axis plotting limits}

\item{ylim}{2-element vector with y-axis plotting limits}

\item{show.quantile.50.90}{If TRUE, 50 (25 to 70) and 90 (5 to 95) percentile regions are plotted, otherwise only the 68 (16 to 84) percentile region is shown. Percentiles are only available if the MF has been fittet using \code{resampling = TRUE} in \code{\link{mffit}}. Otherwise the MF will be displayed with the 68 percent confidence region determined from direct forward-propagation of the parameter covariance matrix.}
}
\value{
If used as \code{mf = mfplot(mf)} (with \code{nbins=NULL} or \code{nbins>0}), the list \code{mf} is appended an additional entry \code{bin}, containing the binned galaxy data points.
}
\description{
This function displays the galaxy mass function (MF) fitted using \code{\link{mffit}}.
}
\examples{
data = mfdata()
mf = mffit(data$x, data$veff, data$sigma, write.fit = FALSE)
mfplot(mf, nbins = 12, bin.xmin = 6.5, bin.xmax = 9.5, xlim=c(2e6,5e10), ylim=c(1e-3,2))

}
\author{
Danail Obreschkow
}
\seealso{
\code{\link{mffit}}
}

