% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfplotveff.R
\name{mfplotveff}
\alias{mfplotveff}
\title{Plot effective volume}
\usage{
mfplotveff(mf)
}
\arguments{
\item{mf}{List produced by \code{\link{mffit}}}
}
\description{
This function plots the function \code{Veff(x)} used for the mass function fit, as stored in the sublist \code{fit} of the output produced by \code{\link{mffit}}.
}
\examples{
data = mfdata()
mf = mffit(data$x, data$veff, data$sigma, write.fit = FALSE)
mfplotveff(mf)

}
\author{
Danail Obreschkow
}
\seealso{
\code{\link{mffit}}
}

